requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-groq>=0.0.1
langchain-huggingface>=0.0.1
langchain-experimental>=0.0.50
langgraph>=0.0.20
langchain-text-splitters>=0.0.1
PyPDF2>=3.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
typing-extensions>=4.8.0
markdown2>=2.4.0
fpdf2>=2.5.0

